fix: address Copilot review findings on AddItem PR
1. Order.Total was hard-coded to 0m with no Items collection at all -
   any future code loading Order via IOrderRepository.GetAsync (not
   just AddItemResult's own separate computation) would see a wrong,
   permanently-zero total, contradicting BR5 and the glossary. Order
   now genuinely owns an Items collection and derives Total from it;
   OrderRepository.GetAsync fetches current items (via
   IOrderItemRepository) when reconstructing, so any caller gets a
   correct aggregate. AddItemResult is built directly from Order
   instead of a separately-computed DTO, removing the duplicate
   total-computation logic entirely.

2. AddItemHandler validated OrderId/MenuItemId existence before
   checking the idempotency claim, so a key reused against a
   different (possibly nonexistent) order/menu item returned 404
   instead of the specified 409 (AC9). The claim is now checked first
   - a mismatch is an immediate conflict regardless of whether the
   request's own identifiers resolve to anything, and a match replays
   without re-validating resources that already proved to exist on
   the original call. The repository's own atomic re-check (for the
   race between this upfront check and a concurrent claim) is
   unchanged.

3. AlcoholicItemLimitPolicy.Exceeds added two ints directly, which can
   overflow (wrap negative) right before the comparison at the
   integer boundary and silently return false instead of rejecting.
   Widened to long arithmetic.

4. Added cross-tenant integration tests for AC2 (order) and AC4 (menu
   item) - the suite only covered "doesn't exist at all", not the
   more security-relevant "exists but belongs to another tenant".

5. The concurrent-merge test verified only the final quantity, not the
   total the same requirement covers - extended the post-race direct
   read to include unit_price_snapshot and assert the derived total.

6/7. The glossary's claim that Order owns items/derives Total is now
   true (see #1). Corrected the PR description, which had called the
   specification "approved" - its Status field is still
   "Draft — Pending Review" on main; that's a human decision (Article
   1.4), not something to self-assign, so it's flagged for the user
   rather than changed here.

77/77 tests passing, stable across 3 repeated runs.

+++ src/Modules/Restaurant/PresenciaVirtual.Modules.Restaurant/Ordering/AddItem/AlcoholicItemLimitPolicy.cs
@@ -4,5 +4,8 @@ namespace PresenciaVirtual.Modules.Restaurant.Ordering.AddItem;
 public static class AlcoholicItemLimitPolicy
 {
     public static bool Exceeds(int existingQuantity, int requestedQuantity, int? maxQuantityPerLine)
-        => maxQuantityPerLine is { } limit && existingQuantity + requestedQuantity > limit;
+        // Widened to long: existingQuantity + requestedQuantity can each be near int.MaxValue,
+        // and adding them as int could overflow (wrapping negative) right before the
+        // comparison, silently returning false instead of correctly rejecting the request.
+        => maxQuantityPerLine is { } limit && (long)existingQuantity + requestedQuantity > limit;
 }
